Remove courses from main, removed cart from footer, changed all prices to decimal

+++ database/migrations/2024_05_09_154232_create_products_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->foreignId('brand_id')->constrained('brands')->onDelete('cascade');
             $table->foreignId('category_id')->constrained('categories')->onDelete('cascade');
             $table->integer('quantity');
-            $table->integer('price');
-            $table->integer('stylist_price');
+            $table->decimal('price');
+            $table->decimal('stylist_price');
 
             $table->timestamps();
         });
